Объединил js-файлы
