+++ package.json
@@ -47,5 +47,8 @@
   ],
   "engines": {
     "node": "8.16"
+  },
+  "dependencies": {
+    "gulp-uglify": "^3.0.2"
   }
 }
