+++ source/catalog.html
@@ -314,11 +314,7 @@
     </form>
   </section>
 
-  <script src="js/picturefill.js"></script>
-  <script src="js/svg4everybody.min.js"></script>
-  <script>svg4everybody();</script>
-  <script src="js/menu.js"></script>
-  <script src="js/modal-catalog.js"></script>
+  <script src="js/script.js"></script>
 
 </body>
