+++ source/form.html
@@ -275,10 +275,7 @@
     </div>
   </footer>
 
-  <script src="js/picturefill.js"></script>
-  <script src="js/svg4everybody.min.js"></script>
-  <script>svg4everybody();</script>
-  <script src="js/menu.js"></script>
+  <script src="js/script.js"></script>
 
 </body>
