+++ source/index.html
@@ -255,7 +255,7 @@
   </main>
   <footer class="page-footer">
     <div class="container">
-      <a class="page-footer__logo">
+      <a class="page-footer__logo page-footer__logo--index">
         <svg class="footer-logo__image" viewBox="0 0 101.615 34.056" width="102" height="34" role="img"
           aria-label="Логотип интернет-магазина 'Мишка'">
           <g>
@@ -331,11 +331,7 @@
     </form>
   </section>
 
-  <script src="js/picturefill.js"></script>
-  <script src="js/svg4everybody.min.js"></script>
-  <script>svg4everybody();</script>
-  <script src="js/menu.js"></script>
-  <script src="js/modal-index.js"></script>
+  <script src="js/script.js"></script>
 
 </body>
