+++ source/sass/blocks/about-us.scss
@@ -101,8 +101,6 @@
     margin-bottom: 25px;
 
     .about-us__title {
-      margin: 0;
-      padding: 0;
       margin-bottom: 47px;
     }
 
@@ -157,6 +155,10 @@
   .about-us {
     .container {
       padding-left: 227px;
+    }
+
+    .about-us__title {
+      margin-left: -5px;
     }
 
     .about-us__item:nth-child(2n+1) {
