+++ source/sass/blocks/contacts.scss
@@ -53,6 +53,7 @@
     width: 100%;
     height: 457px;
     margin-bottom: 34px;
+    border: none;
   }
 
   .contacts__write-button {
