+++ source/sass/blocks/page-footer.scss
@@ -60,6 +60,16 @@
       border-top: none;
     }
   }
+
+  .page-footer__logo:active {
+    fill: #46c1ae;
+    opacity: 0.3;
+  }
+
+  .page-footer__logo--index:active {
+    fill: #000000;
+    opacity: 1;
+  }
 }
 
 @media (min-width: 1150px) {
@@ -71,5 +81,24 @@
     .page-footer__social {
       margin-right: 268px;
     }
+  }
+
+  .page-footer__logo:hover,
+  .page-footer__logo:focus {
+    fill: #46c1ae;
+  }
+
+  .page-footer__logo:active {
+    opacity: 0.3;
+  }
+
+  .page-footer__logo--index:hover,
+  .page-footer__logo--index:focus {
+    fill: #000000;
+  }
+
+  .page-footer__logo--index:active {
+    fill: #000000;
+    opacity: 1;
   }
 }
