+++ source/sass/blocks/section-title.scss
@@ -6,6 +6,7 @@
 
 @media (min-width: 768px) {
   .section-title {
+    margin-left: -5px;
     font-size: 43px;
     line-height: 50px;
   }
@@ -13,6 +14,7 @@
 
 @media (min-width: 1150px) {
   .section-title {
+    margin-left: 0;
     font-size: 53px;
     line-height: 57px;
   }
